requirements.txt
Django==1.5.1
South==0.8.1
dj-database-url==0.2.2
dj-static==0.0.5
django-annoying==0.7.7
django-tastypie==0.9.16
django-toolbelt==0.0.1
gunicorn==17.5
mimeparse==0.1.3
psycopg2==2.5.1
python-dateutil==2.1
python-mimeparse==0.1.4
six==1.3.0
static==0.4
wsgiref==0.1.2
